backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
slowapi
pydantic
pydantic-settings
python-multipart
aiohttp
aiofiles
